[Fix] Circular reference between LikeServiceImpl and PostServiceImpl

+++ src/main/java/com/example/travel_diary/controller/PostController.java
@@ -35,7 +35,7 @@ public class PostController {
     }
 
     @GetMapping("/{id}/likes")
-    public Long getLikes(@PathVariable Long id) {
+    public int getLikes(@PathVariable Long id) {
         return postService.getLikes(id);
     }
 }
